Move benchmark context managers
