+++ src/ggrc/utils/benchmarks.py
@@ -1,0 +1,33 @@
+# Copyright (C) 2016 Google Inc., authors, and contributors
+# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+
+"""Benchmark context managers."""
+
+import time
+
+from flask import current_app
+
+
+class BenchmarkContextManager(object):
+
+  def __init__(self, message):
+    self.message = message
+
+  def __enter__(self):
+    self.start = time.time()
+
+  def __exit__(self, exc_type, exc_value, exc_trace):
+    end = time.time()
+    current_app.logger.info("{:.4f} {}".format(end - self.start, self.message))
+
+
+class WithNop(object):
+
+  def __init__(self, message):
+    pass
+
+  def __enter__(self):
+    pass
+
+  def __exit__(self, exc_type, exc_value, exc_trace):
+    pass
